+++ my_slam/feature_extract/feature_alignment.hpp
@@ -20,7 +20,7 @@ namespace my_slam {
             uint8_t* ref_patch_with_border,
             uint8_t* ref_patch,
             const int n_iter,
-            Eigen::Vector2d& cur_px_estimate,
+            Eigen::Vector2f& cur_px_estimate,
             bool no_simd = false);
 
     bool align2D_SSE2(
